Refactor code and simplify view logic
Move Readme renderer out of product helper module and into own model
ProductReadmeMarkdownRenderer
Create model ProductReadme to handle various Readme needs
Add Product#readme as a getter function for the product's ProductReadme
Split lines in view to improve readability
